+++ .gitignore
@@ -1,3 +1,7 @@
 proxy.jsp
 proxy_er.jsp
 openlayers/
+node_modules/
+.cache
+dist
+package-lock.json
