+++ js/Config.js
@@ -1,0 +1,8 @@
+const PUBLIC_WFS_URL = "proxy.jsp";
+const ER_WFS_URL = "proxy_er.jsp";
+const EPSG_CODE = 'EPSG:25832';
+//const EPSG_CODE = 'EPSG:31467'
+
+exports.PUBLIC_WFS_URL;
+exports.ER_WFS_URL;
+exports.EPSG_CODE;
