+++ js/_old/quer_art_ober.js
