+++ js/_old/quer_config.js
