+++ js/_old/quer_vektor.js
