+++ package.json
@@ -1,0 +1,27 @@
+{
+  "name": "QuerschnittsEditor",
+  "version": "0.0.1",
+  "description": "Ermöglicht das Editieren von Querschnitte im WebEditor",
+  "main": "js/edit.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "parcel *.htm",
+    "build": "parcel build --public-url . index.html"
+  },
+  "author": "Florian Timm",
+  "license": "MIT",
+  "dependencies": {
+    "@babel/polyfill": "^7.2.5",
+    "css": "^2.2.4",
+    "js": "^0.1.0",
+    "ol": "^5.3.0",
+    "proj4": "^2.5.0"
+  },
+  "devDependencies": {
+    "parcel-bundler": "^1.11.0"
+  },
+  "browserslist": [
+    "> 1%",
+    "last 2 versions"
+  ]
+}
